change central client class name
