+++ cli/ballerina-cli/src/main/java/io/ballerina/cli/cmd/PullCommand.java
@@ -23,7 +23,7 @@ import io.ballerina.projects.ProjectException;
 import io.ballerina.projects.SemanticVersion;
 import io.ballerina.projects.util.ProjectConstants;
 import io.ballerina.projects.util.ProjectUtils;
-import org.ballerinalang.central.client.CentralAPIClientV2;
+import org.ballerinalang.central.client.CentralAPIClient;
 import org.ballerinalang.central.client.exceptions.CentralClientException;
 import org.ballerinalang.central.client.exceptions.PackageAlreadyExistsException;
 import org.ballerinalang.toml.model.Settings;
@@ -165,7 +165,7 @@ public class PullCommand implements BLauncherCmd {
             try {
                 Settings settings = readSettings();
                 Proxy proxy = initializeProxy(settings.getProxy());
-                CentralAPIClientV2 client = new CentralAPIClientV2(RepoUtils.getRemoteRepoURL(), proxy);
+                CentralAPIClient client = new CentralAPIClient(RepoUtils.getRemoteRepoURL(), proxy);
                 client.pullPackage(orgName, packageName, version, packagePathInBalaCache, supportedPlatform,
                                    RepoUtils.getBallerinaVersion(), false);
             } catch (PackageAlreadyExistsException e) {
