+++ cli/ballerina-cli/src/main/java/io/ballerina/cli/cmd/SearchCommand.java
@@ -20,7 +20,7 @@ package io.ballerina.cli.cmd;
 
 import io.ballerina.cli.BLauncherCmd;
 import io.ballerina.projects.JvmTarget;
-import org.ballerinalang.central.client.CentralAPIClientV2;
+import org.ballerinalang.central.client.CentralAPIClient;
 import org.ballerinalang.central.client.exceptions.CentralClientException;
 import org.ballerinalang.central.client.model.PackageSearchResult;
 import org.ballerinalang.toml.model.Settings;
@@ -124,7 +124,7 @@ public class SearchCommand implements BLauncherCmd {
         try {
             Settings settings = readSettings();
             Proxy proxy = initializeProxy(settings.getProxy());
-            CentralAPIClientV2 client = new CentralAPIClientV2(RepoUtils.getRemoteRepoURL(), proxy);
+            CentralAPIClient client = new CentralAPIClient(RepoUtils.getRemoteRepoURL(), proxy);
             PackageSearchResult packageSearchResult = client.searchPackage(query,
                                                                            JvmTarget.JAVA_11.code(),
                                                                            RepoUtils.getBallerinaVersion());
